@@ -1,4 +1,4 @@
-name: junagadh_temple
+name: junagadh
 description: A new Flutter project.
 
 # The following line prevents the package from being accidentally published to
@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.15.1 <3.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -38,7 +38,12 @@ dependencies:
   shared_preferences: ^2.0.6
   dio: ^4.0.0
   webview_flutter: ^3.0.0
-
+  package_info_plus: ^1.0.6
+  url_launcher: ^6.0.18
+  path_provider: ^2.0.8
+  intl: ^0.17.0
+  extended_image: ^6.0.1
+  share_plus: ^3.0.4
 
 dev_dependencies:
   flutter_test:
@@ -79,16 +84,7 @@ flutter:
   # list giving the asset and other descriptors for the font. For
   # example:
   fonts:
-  - family: Lato
-    fonts:
-    - asset: assets/fonts/Lato-Regular.ttf
-    - asset: assets/fonts/Lato-Bold.ttf
-  #         style: italic
-  #   - family: Trajan Pro
-  #     fonts:
-  #       - asset: fonts/TrajanPro.ttf
-  #       - asset: fonts/TrajanPro_Bold.ttf
-  #         weight: 700
-  #
-  # For details regarding fonts from package dependencies,
-  # see https://flutter.dev/custom-fonts/#from-packages
+    - family: Lato
+      fonts:
+        - asset: assets/fonts/Lato-Regular.ttf
+        - asset: assets/fonts/Lato-Bold.ttf
