@@ -3,15 +3,16 @@ import 'dart:io';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:junagadh_temple/app/screens/lending_page.dart';
-import 'package:junagadh_temple/app/utils/constants.dart';
-import 'package:junagadh_temple/app/utils/sizer.dart';
+import 'package:junagadh/app/screens/lending_page.dart';
+import 'lib/app/screens/lending_page.dart';
+import 'package:junagadh/app/utils/constants.dart';
+import 'package:junagadh/app/utils/sizer.dart';
 
 Widget app(){
 
   return Sizer(builder: (context, orientation, deviceType) {
     return MaterialApp(
-      title: 'baseStructure',
+      title: 'Junagadh',
       debugShowCheckedModeBanner: false,
       debugShowMaterialGrid: false,
       theme: ThemeData(
@@ -21,12 +22,16 @@ Widget app(){
             color: Colors.white, //change your color here
           ),
             backgroundColor: Colors.white,
+          systemOverlayStyle: SystemUiOverlayStyle(
+            // Status bar color
+            statusBarColor: kPrimaryColor,
+
+          ),
             elevation: 0,
              foregroundColor: kPrimaryColor,
             centerTitle: true,
             backwardsCompatibility: false,
-
-            titleTextStyle: TextStyle(
+          titleTextStyle: TextStyle(
                fontFamily: kRegularFonts,
                 fontSize: 16,
                 fontWeight: FontWeight.bold,
