@@ -1,12 +1,15 @@
 class ServerConfigs {
 
-  static String baseURL =
-      'http://loclpms-api-dev.ap-south-1.elasticbeanstalk.com/Mobile/api/v1/';
+  static String baseURL = 'https://swaminarayanmandirjunagadh.com/wp-json/wp/v2/';
 
-  //static String imageBaseURL = 'http://$domain/';
-  static String userLogin = 'Login/checklogindetails';
-  static String userReservation = 'Login/GetMemberResvationList';
-  static String houseKeeping = 'Member/HouseKeepingSchedule';
-  static String getBankDetail = 'Member/GetMemBankDetailByID';
-  static String userLogout = 'userLogout';
+  //http://swaminarayanmandirjunagadh.softcube.co.in/wp-json/wp/v2/dailydarshan?filter%5Borderby%5D=daily_darshan_date&order=asc&per_page=30
+
+  static String homeSlider = 'homeslider';
+  static String getGallery = 'imagegallery';
+  static String getDailyDarshan = 'dailydarshan';
+  static String getEvents = 'posts';
+  static String getDonation = 'donation';
+  static String getDarshanData = 'dailydarshan/dates';
+  static String getTiming = 'time';
+  static String getYoutube = 'youtube';
 }
