@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:junagadh_temple/app/screens/base_state.dart';
+import 'package:junagadh/app/screens/base_state.dart';
 
 abstract class BaseStateLess extends StatelessWidget with BaseScreen {
   BaseStateLess({Key? key}) : super(key: key);
