@@ -8,7 +8,7 @@ Color bg1 = const Color(0xffD3BA9D);
 Color bg2 = const Color(0xffEBD6BC).withOpacity(0.98);
 Color kLightYellow = const Color(0xffF1D398);
 Color greyColor = const Color(0xff666666);
-Color dividerColor = Color(0xff949494);
+Color dividerColor = const Color(0xff949494);
 
 //Routes
 String kInitialRoute = '/';
@@ -26,29 +26,34 @@ double kFlexibleSize(double size) {
 final kAuthTitleStyle = TextStyle(
   color: kWhite,
   fontSize: 20.sp,
+  fontFamily: kRegularFonts,
   fontWeight: FontWeight.w500,
 );
 
 final kBgButton = TextStyle(
   color: kSecondary,
   fontSize: 16.sp,
+  fontFamily: kRegularFonts,
   fontWeight: FontWeight.w600,
 );
 
 final kImgTitle = TextStyle(
   color: kSecondary,
   fontSize: 16.sp,
+  fontFamily: kRegularFonts,
   fontWeight: FontWeight.w700,
 );
 
 final kDarshanImg = TextStyle(
   color: kSecondary,
   fontSize: 11.sp,
+  fontFamily: kRegularFonts,
   fontWeight: FontWeight.w700,
 );
 
 final kDonateButton = TextStyle(
   color: kWhite,
   fontSize: 14.sp,
+  fontFamily: kRegularFonts,
   fontWeight: FontWeight.w500,
 );
