@@ -12,7 +12,7 @@ final Image bgmaskImage = Image.asset(imagePath + 'bg_mask.png',fit: BoxFit.cont
 final Image maskImage = Image.asset(imagePath + 'mask_img.png',fit: BoxFit.contain);
 final Image designImage = Image.asset(imagePath + 'design.png',fit: BoxFit.contain);
 Image homeImage(Color color) => Image.asset(imagePath + 'home_icon.png',color: color,);
-Image searchImage(Color color) => Image.asset(imagePath + 'search_icon.png',color: color,);
+Image searchImage(Color color) => Image.asset(imagePath + 'darshan-icon.png',color: color,);
 Image galleryImage(Color color) => Image.asset(imagePath + 'gallery_icon.png',color: color,);
 Image eventImage(Color color) => Image.asset(imagePath + 'event_icon.png',color: color,);
 final Image floatingBtn = Image.asset(imagePath + 'floatingbtn.png');
@@ -22,6 +22,10 @@ final Image paytmBtn = Image.asset(imagePath + 'paytm_icon.png');
 final Image morningImage = Image.asset(imagePath + 'morning.png');
 final Image eveningImage = Image.asset(imagePath + 'evening.png');
 final Image sliderImage = Image.asset(imagePath + 'slider_img.png');
+final Image placeholderImg = Image.asset(imagePath + 'placeholder-img.png');
+final Image upiImg = Image.asset(imagePath + 'upi-icon.png');
+final Image tilakImg = Image.asset(imagePath + 'tilak.png');
+
 
 
 
