@@ -1,5 +1,12 @@
 
+import 'dart:io';
+
+import 'package:extended_image/extended_image.dart';
 import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:junagadh/app/utils/constants.dart';
+
+import 'common_images.dart';
 
 class CustomNetWorkImage extends StatelessWidget {
 
@@ -14,21 +21,50 @@ class CustomNetWorkImage extends StatelessWidget {
     try {
 
       final image = NetworkImage(url);
+      
+      return ExtendedImage.network(
+        url,
+        fit: fit,
+        cache: true,
+        loadStateChanged: (state) {
+          switch(state.extendedImageLoadState){
+            case LoadState.loading:
+              // TODO: Handle this case.
+              if(Platform.isAndroid){
+                return Center(child: SizedBox(height: 30,width: 30,child: CircularProgressIndicator(strokeWidth: 3,color: kSecondary,)),);
+              }else{
+                return const Center(child: SizedBox(height: 30,width: 30,child: CupertinoActivityIndicator(animating: true,)),);
+              }
+              break;
+            case LoadState.completed:
+              // TODO: Handle this case.
+              break;
+            case LoadState.failed:
+              // TODO: Handle this case.
+              return Center(child: SizedBox(
+                child: tilakImg,
+              ));
+              break;
+          }
+        },
+        //cancelToken: cancellationToken,
+      );
+      
       return FadeInImage(
         fit: fit ?? BoxFit.cover,
-        placeholder: AssetImage(assetName ?? "assets/images/placeholder.png",),
+        placeholder: AssetImage(assetName ?? "assets/images/placeholder-img.png",),
         image: image,
         imageErrorBuilder: (context, error, stackTrace) {
 
           return Image.asset(
-            assetName ?? "assets/images/placeholder.png",
+            assetName ?? "assets/images/placeholder-img.png",
             fit: fit ?? BoxFit.cover,
           );
         },
         placeholderErrorBuilder: (context, error, stackTrace) {
 
           return Image.asset(
-            assetName ?? "assets/images/placeholder.png",
+            assetName ?? "assets/images/placeholder-img.png",
             fit: fit ?? BoxFit.cover,
           );
         },
